refac: coded_evalutors -> evaluators, associated renames/changes
