+++ src/uipath/eval/evaluators/contains_evaluator.py
@@ -1,6 +1,11 @@
 """Contains evaluator for agent outputs."""
 
-from ..models import AgentExecution, EvaluationResult, EvaluatorType, NumericEvaluationResult
+from ..models import (
+    AgentExecution,
+    EvaluationResult,
+    EvaluatorType,
+    NumericEvaluationResult,
+)
 from .base_evaluator import BaseEvaluationCriteria
 from .output_evaluator import (
     OutputEvaluator,
