+++ src/uipath/eval/evaluators/exact_match_evaluator.py
@@ -1,14 +1,29 @@
-"""Exact match evaluator for binary pass/fail evaluation of agent outputs."""
+"""Exact match evaluator for agent outputs."""
 
-from typing import Any
+from ..models import (
+    AgentExecution,
+    EvaluationResult,
+    EvaluatorType,
+    NumericEvaluationResult,
+)
+from .output_evaluator import (
+    OutputEvaluationCriteria,
+    OutputEvaluator,
+    OutputEvaluatorConfig,
+)
 
-from uipath.eval.models import BooleanEvaluationResult, EvaluationResult
 
-from ..models.models import AgentExecution
-from .deterministic_evaluator_base import DeterministicEvaluatorBase
+class ExactMatchEvaluatorConfig(OutputEvaluatorConfig[OutputEvaluationCriteria]):
+    """Configuration for the exact match evaluator."""
+
+    name: str = "ExactMatchEvaluator"
+    case_sensitive: bool = False
+    negated: bool = False
 
 
-class LegacyExactMatchEvaluator(DeterministicEvaluatorBase[dict[str, Any]]):
+class ExactMatchEvaluator(
+    OutputEvaluator[OutputEvaluationCriteria, ExactMatchEvaluatorConfig, type(None)]  # type: ignore
+):
     """Evaluator that performs exact structural matching between expected and actual outputs.
 
     This evaluator returns True if the actual output exactly matches the expected output
@@ -16,22 +31,38 @@ class LegacyExactMatchEvaluator(DeterministicEvaluatorBase[dict[str, Any]]):
     to floats for consistent comparison.
     """
 
+    @classmethod
+    def get_evaluator_id(cls) -> str:
+        """Get the evaluator id."""
+        return EvaluatorType.EXACT_MATCH.value
+
     async def evaluate(
-        self, agent_execution: AgentExecution, evaluation_criteria: dict[str, Any]
+        self,
+        agent_execution: AgentExecution,
+        evaluation_criteria: OutputEvaluationCriteria,
     ) -> EvaluationResult:
         """Evaluate whether actual output exactly matches expected output.
 
         Args:
             agent_execution: The execution details containing:
                 - agent_input: The input received by the agent
-                - actual_output: The actual output from the agent
-                - spans: The execution spans to use for the evaluation
+                - agent_output: The actual output from the agent
+                - agent_trace: The execution spans to use for the evaluation
             evaluation_criteria: The criteria to evaluate
 
         Returns:
             EvaluationResult: Boolean result indicating exact match (True/False)
         """
-        return BooleanEvaluationResult(
-            score=self._canonical_json(agent_execution.agent_output)
-            == self._canonical_json(evaluation_criteria)
+        actual_output = str(self._get_actual_output(agent_execution))
+        expected_output = str(self._get_expected_output(evaluation_criteria))
+        if not self.evaluator_config.case_sensitive:
+            actual_output = actual_output.lower()
+            expected_output = expected_output.lower()
+
+        is_exact_match = actual_output == expected_output
+        if self.evaluator_config.negated:
+            is_exact_match = not is_exact_match
+
+        return NumericEvaluationResult(
+            score=float(is_exact_match),
         )
