+++ src/uipath/eval/evaluators_types/LLMJudgeStrictJSONSimilarityOutputEvaluator.json
